+++ tinytuya/core.py
@@ -68,7 +68,6 @@
 
 # Modules
 from __future__ import print_function  # python 2.7 support
-from hashlib import md5
 import logging
 import sys
 from colorama import init
@@ -320,43 +319,6 @@ def pad(s):
 
 def unpad(s):
     return s[: -ord(s[len(s) - 1 :])]
-
-def encrypt(msg, key):
-    return AESCipher( key ).encrypt( msg, use_base64=False, pad=True )
-
-def decrypt(msg, key):
-    return AESCipher( key ).decrypt( msg, use_base64=False, decode_text=True )
-
-#def decrypt_gcm(msg, key):
-#    nonce = msg[:12]
-#    return AES.new(key, AES.MODE_GCM, nonce=nonce).decrypt(msg[12:]).decode()
-
-# UDP packet payload decryption - credit to tuya-convert
-udpkey = md5(b"yGAdlopoPVldABfn").digest()
-
-def decrypt_udp(msg):
-    try:
-        header = parse_header(msg)
-    except:
-        header = None
-    if not header:
-        return decrypt(msg, udpkey)
-    if header.prefix == PREFIX_55AA_VALUE:
-        payload = unpack_message(msg).payload
-        try:
-            if payload[:1] == b'{' and payload[-1:] == b'}':
-                return payload.decode()
-        except:
-            pass
-        return decrypt(payload, udpkey)
-    if header.prefix == PREFIX_6699_VALUE:
-        unpacked = unpack_message(msg, hmac_key=udpkey, no_retcode=None)
-        payload = unpacked.payload.decode()
-        # app sometimes has extra bytes at the end
-        while payload[-1] == chr(0):
-            payload = payload[:-1]
-        return payload
-    return decrypt(msg, udpkey)
 
 
 def appenddevice(newdevice, devices):
